Remove subspace-node dep from cirrus-node
`RelayChainCli` will be evolved into `SecondaryChainCli` later.
